move constants to client
