+++ src/server/ClientConnection.test.ts
@@ -3,14 +3,14 @@ import { Socket } from 'socket.io';
 import sinon from 'sinon';
 
 import ClientConnection from './ClientConnection';
-import * as constants from './constants';
+import * as constants from '../client/constants';
 import * as rtpPortIdentifier from './rtpPortIdentifier';
 import Session from './Session';
 import SessionManager from './SessionManager';
 
 let socket;
 
-function identifyNewClient(id, name, sessionId, callback = () => {}) {
+function identifyNewClient(id, name, sessionId, callback = () => { }) {
   const socket = <Socket>new EventEmitter();
   socket.id = id;
   const clientConnection = new ClientConnection(socket);
@@ -37,7 +37,7 @@ afterEach(() => {
 it('should create a new session when the first client identifies itself', () => {
   const clientConnection = new ClientConnection(socket);
 
-  socket.emit('identify', { name: 'Name', sessionId: 'Session' }, () => {});
+  socket.emit('identify', { name: 'Name', sessionId: 'Session' }, () => { });
 
   expect(clientConnection.session).toBeDefined();
   expect(clientConnection.session.size).toBe(1);
@@ -46,7 +46,7 @@ it('should create a new session when the first client identifies itself', () => 
 it('should remember the client properties when a client identifies itself', () => {
   const clientConnection = new ClientConnection(socket);
 
-  socket.emit('identify', { name: 'Name', sessionId: 'Session' }, () => {});
+  socket.emit('identify', { name: 'Name', sessionId: 'Session' }, () => { });
 
   expect(clientConnection.client).toEqual({
     id: 'socket-id',
@@ -76,7 +76,7 @@ it('should announce a new client to the current clients', () => {
   const callback = jest.fn();
   socket1.on('user joined', callback);
 
-  socket.emit('identify', { name: 'Name2', sessionId: 'Session' }, () => {});
+  socket.emit('identify', { name: 'Name2', sessionId: 'Session' }, () => { });
 
   expect(callback).toBeCalled();
   expect(callback.mock.calls[0][0]).toStrictEqual({
